Conflicts:
	engines/configure.engines
	engines/engines.mk
	engines/plugins_table.h

+++ engines/draci/sprite.cpp
@@ -8,12 +8,12 @@
  * modify it under the terms of the GNU General Public License
  * as published by the Free Software Foundation; either version 2
  * of the License, or (at your option) any later version.
-
+ *
  * This program is distributed in the hope that it will be useful,
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
  * GNU General Public License for more details.
-
+ *
  * You should have received a copy of the GNU General Public License
  * along with this program; if not, write to the Free Software
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.
@@ -38,9 +38,9 @@ const Displacement kNoDisplacement = { 0, 0, 1.0, 1.0 };
  *         height height of the image in the buffer
  */
 static void transformToRows(byte *img, uint16 width, uint16 height) {
-	byte *buf = new byte[width * height];
+	byte *buf = new byte[(uint)(width * height)];
 	byte *tmp = buf;
-	memcpy(buf, img, width * height);
+	memcpy(buf, img, (uint)(width * height));
 
 	for (uint16 i = 0; i < width; ++i) {
 		for (uint16 j = 0; j < height; ++j) {
