+++ engines/sword25/gfx/animationtemplate.cpp
@@ -8,12 +8,12 @@
  * modify it under the terms of the GNU General Public License
  * as published by the Free Software Foundation; either version 2
  * of the License, or (at your option) any later version.
-
+ *
  * This program is distributed in the hope that it will be useful,
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
  * GNU General Public License for more details.
-
+ *
  * You should have received a copy of the GNU General Public License
  * along with this program; if not, write to the Free Software
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.
@@ -181,7 +181,7 @@ bool AnimationTemplate::persist(OutputPersistenceBlock &writer) {
 	Result &= AnimationDescription::persist(writer);
 
 	// Frameanzahl schreiben.
-	writer.write(_frames.size());
+	writer.write((uint32)_frames.size());
 
 	// Frames einzeln persistieren.
 	Common::Array<const Frame>::const_iterator Iter = _frames.begin();
@@ -209,7 +209,7 @@ bool AnimationTemplate::unpersist(InputPersistenceBlock &reader) {
 	result &= AnimationDescription::unpersist(reader);
 
 	// Frameanzahl lesen.
-	uint frameCount;
+	uint32 frameCount;
 	reader.read(frameCount);
 
 	// Frames einzeln wieder herstellen.
